% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-images-funs.R
\name{dalle_import}
\alias{dalle_import}
\title{dalle_import function}
\usage{
dalle_import(imgpath)
}
\arguments{
\item{imgpath}{A full path to your image file directory}
}
\value{
Character vector of keywords
}
\description{
Create a keyword list based on image file names in a directory
}
\examples{
dalle_import(imgpath = "~/Pictures/DallE")
}
\keyword{image}
