% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-images-funs.R
\name{dalle_keywords}
\alias{dalle_keywords}
\title{Generate and sort keywords function}
\usage{
dalle_keywords(y)
}
\arguments{
\item{y}{A tibble}
}
\value{
Keywords
}
\description{
Generate and sort keywords function
}
\examples{
dalle_keywords()
}
