% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imgsortR-package.R
\docType{package}
\name{imgsortR-package}
\alias{imgsortR}
\alias{imgsortR-package}
\title{imgsortR: Sort Directory of Images from Dall-E 2 into Folders Based on Keywords}
\description{
This package works based on the assumption that your Dall E 2 images have prompt text as the file name. It then extracts a keyword list, and attempts to sort your images into new (or existing) folders based on the keywords. Sorting happens from the highest frequency keyword to the lowest frequency. This means that an image can be sorted only into one folder, being the highest frequency keyword that has a match in the file name.
}
\keyword{internal}
