% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort-images-funs.R
\name{sort_images}
\alias{sort_images}
\title{Sort images function}
\usage{
sort_images(keywords, imgpath)
}
\arguments{
\item{keywords}{Character vector of keywords}

\item{imgpath}{A full path to your image file directory}
}
\value{
Sorts images
}
\description{
This uses a keyword list generated by dalle_import and sorts images into folders.
}
\examples{
sort_images(keywords = my-key-words, imgpath = "~/Pictures/DallE")
}
